Moved the hub class to server/bsp
